# feat: cache JSON parsing in datafile lookups