## Background

When building datafiles, Featurevisor stringifies segments in rules and their conditions by default: https://featurevisor.com/docs/configuration/#stringify

When JavaScript SDK evaluates a feature, it parses relevant stringified JSON just in time, every time.

## Issue

It's possible that this just in time JSON parsing for every evaluation (even if repeated using same SDK instance), can lead to additional time that we can avoid.

TODO: prove with real numbers.

## What's done

When a segment/condition stringified JSON is about to be parsed, it will now check if it has been parsed before and return the cached parsed value.

## Benchmarking

Via `example-1` project:

### Without optimization (main branch)

```
$ npx featurevisor benchmark --environment=production --feature=test --context='{"userId": "fh", "version": "5.0.1"}' -n=1000000

Running benchmark for feature "test"...

Building datafile containing all features for "production"...
Datafile build duration: 11ms
Datafile size: 9.20 kB

...SDK initialized

Against context: {"userId":"fh","version":"5.0.1"}
Evaluating flag 1000000 times...

Evaluated value : true
Total duration  : 2s 53ms
Average duration: 0.002053ms
```

### With optimization (this PR)

```
$ npx featurevisor benchmark --environment=production --feature=test --context='{"userId": "fh", "version": "5.0.1"}' -n=1000000

Running benchmark for feature "test"...

Building datafile containing all features for "production"...
Datafile build duration: 10ms
Datafile size: 9.20 kB

...SDK initialized

Against context: {"userId":"fh","version":"5.0.1"}
Evaluating flag 1000000 times...

Evaluated value : true
Total duration  : 2s 371ms
Average duration: 0.002371ms
```

### Results

No visible improvement in speed.

## TODOs

- [x] Implement caching in DatafileReader
- [ ] Make `DatafileReader` a bit more DRY
- [ ] Get rid of just in time parsing outside of DatafileReader
- [ ] Consider adding a new option for enabling caching for JSON parsing (defaulting to false for now may be)
- [x] Run benchmarks/tests with and without this optimization
- [x] Write PR description
- [ ] Merge if numbers are good